requirements_dev.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-bootstrap4==22.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-widget-tweaks==1.4.12
idna==3.3
mysqlclient==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
